add unique person id

+++ cpp/models/abm/person_id.h
@@ -30,16 +30,16 @@ namespace abm
 {
 
 /// Unique identifier for a Person within a Model.
-struct MEMILIO_ENABLE_EBO PersonId : public mio::TypeSafe<uint64_t, PersonId>, public OperatorComparison<PersonId> {
+struct MEMILIO_ENABLE_EBO PersonId : public mio::TypeSafe<uint32_t, PersonId>, public OperatorComparison<PersonId> {
     /// @brief Create an ID.
-    PersonId(uint64_t id)
-        : mio::TypeSafe<uint64_t, PersonId>(id)
+    PersonId(uint32_t id)
+        : mio::TypeSafe<uint32_t, PersonId>(id)
     {
     }
 
     /// @brief Create an invalid ID.
     PersonId()
-        : mio::TypeSafe<uint64_t, PersonId>(std::numeric_limits<uint32_t>::max())
+        : mio::TypeSafe<uint32_t, PersonId>(std::numeric_limits<uint32_t>::max())
     {
     }
